// web/default/src/features/home/types.ts
/*
Copyright (C) 2023-2026 QuantumNous

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU Affero General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version.
*/

export interface HomePageContentResponse {
  success: boolean
  message?: string
  data?: string
}

export interface HomePageContentResult {
  content: string
  isLoaded: boolean
  isUrl: boolean
}

export interface ShowcaseAsset {
  id: string
  image: string
  title: string
  tags: string
  aspectRatio?: number
}